♻️Refactor: 캐릭터 에셋 폴더 생성
